refactor(smallGroup): vo 디렉토리 위치 변경 - domain 내부로 이동
