+++ src/test/java/com/example/wegather/group/domain/repotitory/SmallGroupRepositoryTest.java
@@ -7,7 +7,7 @@ import com.example.wegather.global.vo.Address;
 import com.example.wegather.global.vo.MemberType;
 import com.example.wegather.global.vo.PhoneNumber;
 import com.example.wegather.group.domain.SmallGroup;
-import com.example.wegather.group.vo.MaxMemberCount;
+import com.example.wegather.group.domain.vo.MaxMemberCount;
 import com.example.wegather.interest.domain.Interests;
 import com.example.wegather.member.domain.Member;
 import com.example.wegather.member.domain.vo.Password;
